requirements.txt
einops
jiwer
Levenshtein
loguru
matplotlib
numpy
PyYAML
scipy
thop
timm
torch